Improve Grouping of Nodes
Instead of adding the new complex node multiple times to different parent elements we now search for the selected node with the lowest depth to determine in which node the new complex will be inserted in.

Therefore we added a breadth-first-search method to CompositeComplex.

Also we now store the RootNode of our graph which makes it easier to check if someone tries to group the root node (which is not allowed).